refactor: label페이지 관련 컴포넌트 이동
- label페이지의 하위 컴포넌트들 중 중복되지않는 컴포넌트들을 label페이지의 하위 폴더로 이동시켰다
