+++ client/src/pages/Labels/EditLabel/index.jsx
@@ -1,5 +1,5 @@
 /* eslint-disable react/button-has-type */
-import React, { useState } from 'react';
+import React, { useState, useRef } from 'react';
 import {
   FlexDiv,
   LabelIcon,
@@ -12,8 +12,20 @@ import {
 
 const EditLabel = ({ action, label }) => {
   const [nameValue, setNameValue] = useState(label.title);
-  const [desValue, setDesValue] = useState(label.content);
   const [colorValue, setColorValue] = useState(label.color);
+  const desValue = useRef(null);
+  const isDisabled =
+    !nameValue || !(colorValue.length === 7 || colorValue.length === 4);
+
+  const changeHandler = (e) => {
+    if (e.target.value === '') {
+      e.target.value = '#';
+    }
+    if (e.target.value.length > 7) {
+      e.target.value = e.target.value.substring(0, 7);
+    }
+    setColorValue(e.target.value);
+  };
 
   return (
     <Container>
@@ -33,9 +45,9 @@ const EditLabel = ({ action, label }) => {
         <Div>
           <div>Description</div>
           <Input
-            value={desValue}
+            defaultValue={label.content}
+            ref={desValue}
             placeholder="Description (optional)"
-            onChange={(e) => setDesValue(e.target.value)}
           />
         </Div>
         <Div>
@@ -44,28 +56,13 @@ const EditLabel = ({ action, label }) => {
             <Refresh color={colorValue} />
             <ColorInput
               value={colorValue}
-              onChange={(e) => {
-                if (e.target.value === '') {
-                  e.target.value = '#';
-                }
-                if (e.target.value.length > 7) {
-                  e.target.value = e.target.value.substring(0, 7);
-                }
-                setColorValue(e.target.value);
-              }}
+              onChange={(event) => changeHandler(event)}
             />
           </FlexDiv>
         </Div>
         <div>
           <button onClick={() => action(false)}>cancel</button>
-          <button
-            disabled={
-              !nameValue ||
-              !(colorValue.length === 7 || colorValue.length === 4)
-            }
-          >
-            Save changes
-          </button>
+          <button disabled={isDisabled}>Save changes</button>
         </div>
       </FlexDiv>
     </Container>
