+++ client/src/pages/Labels/LabelList/index.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react';
 import LabelItem from '../LabelItem';
-import { getLabelsAPI } from '../../apis/label';
+import { getLabelsAPI } from '../../../apis/label';
 import { Container, Header } from './styled';
 
 const LabelList = () => {
@@ -10,11 +10,12 @@ const LabelList = () => {
     const result = await getLabelsAPI();
     setLabelList(result);
   }, []);
+
   return (
     <Container>
       <Header>{labelList.length} labels</Header>
-      {labelList?.map((item) => (
-        <LabelItem label={item} />
+      {labelList?.map((item, index) => (
+        <LabelItem key={index} label={item} />
       ))}
     </Container>
   );
